+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Validation when `api.timing.update()` updated the current active timing against old values
+
 ## [1.28.0] - 2025-07-07
 ### Added
 - New applet commands for file system functions - `wipeout()`, `link()`, `createArchive()`, aHash screenshots - `takeAndUploadScreenshot()`, scheduled power actions - `setScheduledReboot()` etc., device & organization tags - `getOrganizationTags()`